[bugfix] syntax error and strang follow-up error
`install.yml` had a syntax error where

```
when: "{{ item.name }} is not defined"
```

was written instead of:

```
when: item.name is not defined
```

This led to the `postgresql.service` not starting on `@mattermost01`. Apparently
the path to `pg_ctl` wasn't specified correctly. (I don't know why the service
was working before.)

The fix for that follow-up issue involved accessing the postgresql version
(`pg_ver.stdout`) because the correct path is version-dependent.